Add support for multi line messages
When writing a log message, pressing F2 (temporary mapping) will
continue the message onto the next line. The next time Enter is pressed,
a timestamp will be added to the first line that F2 was pressed on, then
return the cursor to the previous line and insert a newline.

Note on the mapping chosen: Ideally this would be Shift+Enter, since
that is much more intuitive, but unfortunately terminals usually don't
distinguish between Shift+Enter and Enter.